backend/requirements.txt
Django>=4.1.3, <4.2
django-admin-interface>=0.23.0, <0.24
djangorestframework>=3.14.0, <3.15
django-cors-headers>=3.13.0, <3.15.0
psycopg2-binary>=2.9.5, <2.10
gunicorn>=20.1.0